Refactoring create email for mobile
- Renamed SignUpPage to CreateEmail for some more clarity
- Split out mobile/desktop UI + styled mobile accordingly
