+++ packages/web/src/pages/sign-up-page/pages/CreateEmailPage.module.css
@@ -1,0 +1,49 @@
+.logo {
+  height: 100%;
+  width: 100%;
+  object-fit: contain;
+}
+.logo.desktop {
+  max-height: 160px;
+  max-width: 160px;
+}
+
+.logo.mobile {
+  max-height: 50px;
+  max-width: 180px;
+}
+
+.mobileCurvedContainer {
+  border-bottom-left-radius: var(--harmony-border-radius-2xl);
+  border-bottom-right-radius: var(--harmony-border-radius-2xl);
+}
+
+.backgroundImageRoot {
+  background: radial-gradient(
+      77.16% 77.16% at 50% 51.81%,
+      rgba(91, 35, 225, 0.8) 0%,
+      rgba(113, 41, 230, 0.64) 67.96%,
+      rgba(162, 47, 235, 0.5) 100%
+    ),
+    url('../../../assets/img/2-DJ-4-3.jpg'), lightgray 50% / cover no-repeat;
+  background-position-y: bottom;
+  background-position-x: -110px;
+  background-size: 150%;
+}
+
+/* TODO: consider adding these to a shared spot, maybe Harmony? */
+.flex1 {
+  flex: 1;
+}
+
+.w100 {
+  width: 100%;
+}
+
+.h100 {
+  height: 100%;
+}
+
+.textAlignCenter {
+  text-align: center;
+}
