--- packages/web/src/pages/sign-up-page/pages/SignUpPage.module.css
@@ -1,15 +1,0 @@
-.logo {
-  max-height: 160px;
-  max-width: 160px;
-  height: 100%;
-  width: 100%;
-  object-fit: contain;
-}
-
-.flex1 {
-  flex: 1;
-}
-
-.w100 {
-  width: 100%;
-}
